HTML scanner and TagAction down to 0, 180 total

+++ MPWHTMLFileTextExtractor.h
@@ -17,5 +17,5 @@
 	id	metadata;
 }
 
-+extractTextFromPath:url maxLength:(int)maxLen attributes:dict;
++extractTextFromPath:url maxLength:(long)maxLen attributes:dict;
 @end
